feat: add profile and settings

+++ src/actions/blog/likeBlog.ts
@@ -6,8 +6,14 @@ import { client } from "@/lib/prismaClient"
 
 export async function likeBlog(blogId: string) {
     try {
+        if (!blogId) return sendError("Invalid Request")
         const { user } = await auth.getCurrentUser()
         if (!user) return sendError("User not logged In")
+
+        const blogExists = await client.blogs.findFirst({ where: { id: blogId } })
+        if (!blogExists) return sendError("Invalid Request")
+
+        // deleting like if it already exists
         const likeExists = await client.likes.findFirst({ where: { blogId, userId: user.id } })
         if (likeExists) {
             await client.likes.deleteMany({ where: { blogId, userId: user.id } })
